+++ modules/azure/aks/variables.tf
@@ -50,6 +50,20 @@ variable "aks_config" {
 
   validation {
     condition = alltrue([
+      for np in concat(var.aks_config.node_pools, [{ version : var.aks_config.version }]) : can(regex("^1.(20|21|22)", np.version))
+    ])
+    error_message = "The Kubernetes version has not been validated yet, supported versions are 1.20, 1.21, 1.22."
+  }
+
+  validation {
+    condition = alltrue([
+      for np in var.aks_config.node_pools : split(".", np.version)[1] <= split(".", var.aks_config.version)[1] && split(".", np.version)[2] <= split(".", var.aks_config.version)[2]
+    ])
+    error_message = "The node Kubernetes version should not be newer than the cluster version, upgrade the cluster first."
+  }
+
+  validation {
+    condition = alltrue([
       for np in var.aks_config.node_pools : length(np.name) <= 12
     ])
     error_message = "The name value cannot be longer than 12 characters."